Add portable fallback to SConstruct when Qt tools are not on PATH
Previous method of finding important Python locations relied on
finding Qt tools first and then estimating relevant Python
directories relative to tools directory. This has some issues:

* Logic for finding package directory was Windows specific.
* Qt tools won't be found unless the scripts directory is already
  on PATH (by activating the virtual environment, for example).
  As a result, the build cannot be triggered via virtual
  environment's Python executable.

The new method searches for Qt tools on external environment's PATH
and invokes them as full paths instead of appending to construction
environment's PATH. This is useful when debugging as one can see
exactly which tool was invoked. As a fallback when Qt tools are not
on PATH, sysconfig is used to guess the path to the scripts
directory in a portable manner. This allows build to try proceeding
even when virtual environment isn't activated.