Add resources as explicit dependencies in build script
Fixes incorrect sdist builds that would omit all resources, simply
because dependencies are implicit in the inclusion of .qrc file.
To include the resources listed in .qrc file, the resulting _rc.py
file must explicitly depend on these resources in addition to the
.qrc file. Using a Scanner doesn't help because it creates implicit
dependencies that are not caught by FindSourceFiles().